feat: Compliance history
