+++ frontend/src/app/components/containers/parties/view/party-view.tsx
@@ -1,31 +1,17 @@
-import { FC, useCallback } from "react";
+import { FC } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import { PartyHeader } from "./party-header";
 import { PartyTabs } from "./party-tabs";
 import { PartyHistoryTab } from "./party-history-tab";
 import { useGraphQLQuery } from "@graphql/hooks";
 import { gql } from "graphql-request";
-import {
-  Party,
-  Investigation,
-  Inspection,
-  CaseFile,
-  InspectionParty,
-  InvestigationParty,
-  Person,
-  ContactMethod,
-  Address,
-  Contravention,
-} from "@/generated/graphql";
+import { Party, Person, ContactMethod, Address } from "@/generated/graphql";
 import { Badge, Button } from "react-bootstrap";
 import { PartyCarousel } from "@/app/components/containers/parties/attachments/party-carousel";
-import { CaseActivities } from "@/app/constants/case-activities";
 import { PartyTypes } from "@/app/constants/party-types";
 import {
-  selectAgencyDropdown,
   selectApproximateAgeDropdown,
   selectBuildDropdown,
-  selectCodeTable,
   selectComplexionDropdown,
   selectEyeColourDropdown,
   selectFacialHairStyleDropdown,
@@ -35,38 +21,13 @@ import {
 } from "@/app/store/reducers/code-table";
 import { useAppSelector } from "@/app/hooks/hooks";
 import Option from "@apptypes/app/option";
-import { CODE_TABLE_TYPES } from "@/app/constants/code-table-types";
-import { CASE_ACTIVITY_TYPES } from "@/app/constants/case-activity-types";
 import { formatPhoneNumber } from "react-phone-number-input/input";
 import { calculateAgeYears, formatDateOfBirth, isYoungPerson } from "@common/methods";
 import { ContactMethods } from "@/app/constants/contact-methods";
-import { getUserAgency } from "@/app/service/user-service";
 import { selectCountries, selectCountrySubdivisions } from "@/app/store/reducers/code-table-selectors";
 import { cmToFeetInches, kgToLb } from "@/app/components/containers/parties/form/party-form-utils";
 import { BUSINESS_IDENTIFIER_LABELS } from "@/app/constants/business-identifiers";
-import { selectOfficers } from "@/app/store/reducers/officer";
-import { useLegislation } from "@/app/graphql/hooks/useLegislationSearchQuery";
-import PartyComplianceHistory from "@/app/components/containers/parties/view/party-compliance-history";
-
-type PartyRelation = {
-  caseId?: string | null;
-  caseName?: string | null;
-  activities?: PartyActivity[] | null;
-  leadAgency?: string | null;
-  sameAgency?: boolean;
-};
-
-type PartyActivity = {
-  id?: string | null;
-  name?: string | null;
-  activityType?: string | null;
-  leadAgency?: string | null;
-  role?: string | null;
-  sameAgency?: boolean;
-  status: string;
-  primaryInvestigatorName?: string;
-  contraventions?: Contravention[] | null;
-};
+import { PartyComplianceHistory } from "@/app/components/containers/parties/view/compliance-history/party-compliance-history";
 
 export const GET_PARTY = gql`
   query GetParty($partyIdentifier: String!) {
@@ -160,90 +121,6 @@ export const GET_PARTY = gql`
           }
         }
       }
-    }
-  }
-`;
-
-const GET_INVESTIGATIONS_BY_PARTY = gql`
-  query GetInvestigationsByParty($partyId: String!, $partyType: String!) {
-    partyHistoryInvestigations(partyId: $partyId, partyType: $partyType) {
-      investigationGuid
-      description
-      leadAgency
-      openedTimestamp
-      updatedTimestamp
-      locationGeometry
-      locationAddress
-      locationDescription
-      name
-      investigationStatus {
-        shortDescription
-      }
-      primaryInvestigatorGuid
-      contraventions {
-        contraventionIdentifier
-        legislationIdentifierRef
-        investigationParty {
-          partyReference
-          enforcementActions {
-            enforcementActionIdentifier
-            enforcementActionCode {
-              enforcementActionCode
-            }
-          }
-        }
-      }
-    }
-  }
-`;
-
-const GET_INSPECTIONS_BY_PARTY = gql`
-  query GetInspectionsByParty($partyId: String!, $partyType: String!) {
-    partyHistoryInspections(partyId: $partyId, partyType: $partyType) {
-      inspectionGuid
-      description
-      leadAgency
-      openedTimestamp
-      updatedTimestamp
-      locationGeometry
-      locationAddress
-      locationDescription
-      name
-    }
-  }
-`;
-
-const GET_CASE_FILES_BY_ACTIVITIES = gql`
-  query GetCaseFilesByActivityIds($activityIdentifiers: [String!]!) {
-    partyHistoryCaseFiles(activityIdentifiers: $activityIdentifiers) {
-      caseIdentifier
-      name
-      leadAgency {
-        agencyCode
-      }
-      activities {
-        activityIdentifier
-      }
-    }
-  }
-`;
-
-const GET_INSPECTION_PARTY_ROLES = gql`
-  query GetInspectionPartyByReference($partyRefId: String!) {
-    InspectionParties(partyRefId: $partyRefId) {
-      inspectionGuid
-      partyReference
-      partyAssociationRole
-    }
-  }
-`;
-
-const GET_INVESTIGATION_PARTY_ROLES = gql`
-  query GetInvestigationPartyByReference($partyRefId: String!) {
-    InvestigationParties(partyRefId: $partyRefId) {
-      investigationGuid
-      partyReference
-      partyAssociationRole
     }
   }
 `;
@@ -380,8 +257,6 @@ export const PartyView: FC = () => {
   const { id = "", tabKey } = useParams<PartyParams>();
   const navigate = useNavigate();
   const currentTab = tabKey || "details";
-  const leadAgencyOptions = useAppSelector(selectAgencyDropdown);
-  const partyRoles = useAppSelector(selectCodeTable(CODE_TABLE_TYPES.PARTY_ASSOCIATION_ROLE));
   const genderOptions = useAppSelector(selectGenderDropdown);
   const approximateAgeOptions = useAppSelector(selectApproximateAgeDropdown);
   const countryOptions = useAppSelector(selectCountries);
@@ -436,186 +311,7 @@ export const PartyView: FC = () => {
   const imperialHeight = cmToFeetInches(partyData?.person?.heightInCm ?? 0);
   const imperialWeight = kgToLb(partyData?.person?.weightInKg ?? 0);
 
-  const getPartyRoleText = (roleCode: string, activityType: string) => {
-    const partyRoleText: string = partyRoles.find(
-      (partyRole) => partyRole.partyAssociationRole === roleCode && partyRole.caseActivityTypeCode === activityType,
-    )?.shortDescription;
-    return partyRoleText;
-  };
-
-  const officers = useAppSelector(selectOfficers);
-
-  const getOfficerName = useCallback(
-    (officerGuid: string): string => {
-      const officer = officers?.find((o) => o.app_user_guid === officerGuid);
-      return officer ? `${officer.last_name}, ${officer.first_name}` : "-";
-    },
-    [officers],
-  );
-
-  const GetPartyRelations = (
-    partyId: string,
-    partyType: string,
-  ): {
-    partyRelations: PartyRelation[];
-    isPartyRelationLoading: boolean;
-  } => {
-    const { relatedInvestigations, relatedInspections, isActivityLoading } = GetRelatedActivities(
-      partyId ?? "",
-      partyType ?? "",
-    );
-
-    const partyRelations = [];
-
-    const userAgency = getUserAgency();
-
-    const relatedInvestigationsIds = relatedInvestigations?.map((investigation) => investigation.investigationGuid);
-    const relatedInspectionIds = relatedInspections?.map((inspection) => inspection.inspectionGuid);
-
-    const activityIds = [...(relatedInvestigationsIds ?? []), ...(relatedInspectionIds ?? [])];
-
-    const { data: caseData, isLoading: isCaseLoading } = useGraphQLQuery<{
-      partyHistoryCaseFiles: CaseFile[];
-    }>(GET_CASE_FILES_BY_ACTIVITIES, {
-      queryKey: ["activities", activityIds],
-      variables: { activityIdentifiers: activityIds },
-      enabled: !!activityIds && activityIds.length > 0,
-    });
-
-    const relatedCases = caseData?.partyHistoryCaseFiles;
-
-    const uniqueCaseIds = [...new Set(relatedCases?.map((item) => item.caseIdentifier))];
-
-    const { rolesInInvestigations, rolesInInspections, isPartyRoleLoading } = GetPartyRoles(id);
-
-    for (let uniqueCaseId of uniqueCaseIds) {
-      const partyRelation: PartyRelation = {};
-      const currentCase = relatedCases?.find((relatedCase) => relatedCase.caseIdentifier === uniqueCaseId);
-      partyRelation.caseId = uniqueCaseId;
-      partyRelation.caseName = currentCase?.name;
-      partyRelation.activities = [];
-      partyRelation.leadAgency = currentCase?.leadAgency?.agencyCode;
-      partyRelation.sameAgency = currentCase?.leadAgency?.agencyCode === userAgency;
-
-      for (let caseActivity of currentCase?.activities ?? []) {
-        const currenInvestigation = relatedInvestigations?.find(
-          (relatedInvestigation) => relatedInvestigation.investigationGuid === caseActivity?.activityIdentifier,
-        );
-
-        if (currenInvestigation) {
-          partyRelation.activities.push({
-            id: currenInvestigation.investigationGuid,
-            name: currenInvestigation.name,
-            activityType: CaseActivities.INVESTIGATION,
-            leadAgency: currenInvestigation?.leadAgency,
-            role: getPartyRoleText(
-              rolesInInvestigations.find(
-                (investigation) => investigation.investigationGuid === currenInvestigation.investigationGuid,
-              )?.partyAssociationRole ?? "",
-              CASE_ACTIVITY_TYPES.INVESTIGATION,
-            ),
-            sameAgency: currenInvestigation?.leadAgency === userAgency,
-            status: currenInvestigation?.investigationStatus?.shortDescription ?? "",
-            primaryInvestigatorName: getOfficerName(currenInvestigation?.primaryInvestigatorGuid ?? ""),
-            contraventions: (currenInvestigation?.contraventions as Contravention[]) ?? null,
-          });
-        }
-
-        const currenInspection = relatedInspections?.find(
-          (relatedInspection) => relatedInspection.inspectionGuid === caseActivity?.activityIdentifier,
-        );
-        if (currenInspection) {
-          partyRelation.activities.push({
-            id: currenInspection.inspectionGuid,
-            name: currenInspection.name,
-            activityType: CaseActivities.INSPECTION,
-            leadAgency: leadAgencyOptions.find((option: Option) => option.value === currenInspection?.leadAgency)
-              ?.label,
-            role: getPartyRoleText(
-              rolesInInspections.find((inspection) => inspection.inspectionGuid === currenInspection.inspectionGuid)
-                ?.partyAssociationRole ?? "",
-              CASE_ACTIVITY_TYPES.INSPECTION,
-            ),
-            sameAgency: currenInspection?.leadAgency === userAgency,
-            status: currenInspection?.inspectionStatus?.shortDescription ?? "",
-          });
-        }
-      }
-      partyRelations.push(partyRelation);
-    }
-    const isPartyRelationLoading = isActivityLoading || isCaseLoading || isPartyRoleLoading;
-
-    return { partyRelations, isPartyRelationLoading };
-  };
-
-  const GetRelatedActivities = (
-    partyId: string,
-    partyType: string,
-  ): {
-    relatedInvestigations: Investigation[];
-    relatedInspections: Inspection[];
-    isActivityLoading: boolean;
-  } => {
-    const { data: investigationData, isLoading: isInvestigationLoading } = useGraphQLQuery<{
-      partyHistoryInvestigations: Investigation[];
-    }>(GET_INVESTIGATIONS_BY_PARTY, {
-      queryKey: ["InvestigationParty", id],
-      variables: { partyId, partyType },
-      enabled: !!partyId && !!partyType,
-    });
-
-    const relatedInvestigations = investigationData?.partyHistoryInvestigations ?? [];
-
-    const { data: inspectionData, isLoading: isInspectionLoading } = useGraphQLQuery<{
-      partyHistoryInspections: Inspection[];
-    }>(GET_INSPECTIONS_BY_PARTY, {
-      queryKey: ["InspectionParty", id],
-      variables: { partyId, partyType },
-      enabled: !!partyId && !!partyType,
-    });
-
-    const relatedInspections = inspectionData?.partyHistoryInspections ?? [];
-
-    const isActivityLoading = isInspectionLoading || isInvestigationLoading;
-
-    return { relatedInvestigations, relatedInspections, isActivityLoading };
-  };
-
-  const GetPartyRoles = (
-    id: string,
-  ): {
-    rolesInInvestigations: InvestigationParty[];
-    rolesInInspections: InspectionParty[];
-    isPartyRoleLoading: boolean;
-  } => {
-    const { data: rolesInInvestigationsData, isLoading: isInvestigationRoleLoading } = useGraphQLQuery<{
-      InvestigationParties: InvestigationParty[];
-    }>(GET_INVESTIGATION_PARTY_ROLES, {
-      queryKey: ["InvestigationPartyRoles", id],
-      variables: { partyRefId: id },
-      enabled: !!id,
-    });
-
-    const rolesInInvestigations = rolesInInvestigationsData?.InvestigationParties ?? [];
-
-    const { data: rolesInInspectionsData, isLoading: isInspectionRoleLoading } = useGraphQLQuery<{
-      InspectionParties: InspectionParty[];
-    }>(GET_INSPECTION_PARTY_ROLES, {
-      queryKey: ["InspectionPartyRoles", id],
-      variables: { partyRefId: id },
-      enabled: !!id,
-    });
-
-    const rolesInInspections = rolesInInspectionsData?.InspectionParties ?? [];
-
-    const isPartyRoleLoading = isInvestigationRoleLoading || isInspectionRoleLoading;
-
-    return { rolesInInvestigations, rolesInInspections, isPartyRoleLoading };
-  };
-
-  const { partyRelations, isPartyRelationLoading } = GetPartyRelations(partyId ?? "", partyType ?? "");
-
-  if (isLoading || isPartyRelationLoading) {
+  if (isLoading) {
     return (
       <div className="comp-complaint-details">
         <PartyHeader />
@@ -718,12 +414,13 @@ export const PartyView: FC = () => {
                 )}
               </div>
               <br />
-              {partyRelations && partyRelations.length > 0 && (
-                <PartyComplianceHistory
-                  partyRelations={partyRelations}
-                  id={id}
-                />
-              )}
+
+              <PartyComplianceHistory
+                partyReference={id}
+                partyTypeGuid={partyId ?? ""}
+                partyType={partyType ?? ""}
+              />
+
               <br />
               <h4>Contact information</h4>
